change private

+++ src/Model.cpp
@@ -15,8 +15,8 @@ void Model::setMapSize(int x_grid_num, int y_grid_num)
   assert(x_grid_num > 0 && y_grid_num > 0);
 
   _grid_map.init(x_grid_num, y_grid_num);
-  for (int i = 0; i < _grid_map.get_x_grid_num(); i++) {
-    for (int j = 0; j < _grid_map.get_y_grid_num(); j++) {
+  for (int i = 0; i < (int) _grid_map.get_x_grid_num(); i++) {
+    for (int j = 0; j < (int) _grid_map.get_y_grid_num(); j++) {
       _grid_map[i][j].set_coord(i, j);
     }
   }
@@ -24,8 +24,10 @@ void Model::setMapSize(int x_grid_num, int y_grid_num)
 
 Node* Model::setNode(const Coordinate& coord, const NodeType& node_type)
 {
-  assert(0 <= coord.get_x() && coord.get_x() < _grid_map.get_x_grid_num());
-  assert(0 <= coord.get_y() && coord.get_y() < _grid_map.get_y_grid_num());
+  assert(0 <= coord.get_x()
+         && coord.get_x() < (int) _grid_map.get_x_grid_num());
+  assert(0 <= coord.get_y()
+         && coord.get_y() < (int) _grid_map.get_y_grid_num());
 
   Node& node = _grid_map[coord.get_x()][coord.get_y()];
   if (node.get_type() == NodeType::kNone) {
@@ -110,7 +112,7 @@ void Model::printGridMap()
 {
   std::cout << "--------------------------------------------------\n";
   for (int j = _grid_map.get_y_grid_num() - 1; j >= 0; j--) {
-    for (int i = 0; i < _grid_map.get_x_grid_num(); i++) {
+    for (int i = 0; i < (int) _grid_map.get_x_grid_num(); i++) {
       printNode(_grid_map[i][j]);
     }
     std::cout << "\n";
